Remove YamlEqual metho to remove duplicate work and bash commands (#65)
Description of changes:
Remove yaml equality check from `util.go` as it is not necessary if we check the spec like we check the status. This will also remove the usage of bash commands such as `sed`, `diff`, and `cat`.
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.